requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Login==0.5.0
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
Werkzeug==1.0.1
